+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Version 0.9.0 (unreleased)
 
+**Fixes**
+
+- Fixed a bug with the parsing of JSON Pointers. When given an arbitrary string without slashes, `JSONPointer` would resolve to the document root. The empty string is the only valid pointer that should resolve to the document root. See [#27](https://github.com/jg-rp/python-jsonpath/issues/27).
+
 **Features**
 
 - Added a command line interface, exposing JSONPath, JSON Pointer and JSON Patch features.